requirements.txt
streamlit
pandas
plotly
openpyxl
openai
python-dotenv
playwright
requests
